Further split into into downloader and info provider
